0x13. C - More singly linked lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include <stdio.h>
 
 /**
  * print_listint - prints all the elements of a linked list
@@ -10,7 +11,7 @@
 size_t print_listint(const listint_t *h)
 {
 	const listint_t *cursor = h;
-	size_t count = 0;
+	int count = 0;
 
 	while (cursor != NULL)
 	{
